requirements.txt
mcp-server
aiohttp
python-dotenv
typing-extensions
rich
fastmcp
requests
rich
pyyaml
capstone
pycryptodome